estimateMode {fabia}	R Documentation

Estimation of the modes of the rows of a matrix

Description

estimateMode: R implementation of estimateMode.

Usage

estimateMode(X,maxiter=50,tol=0.001,alpha=0.1,a1=4.0,G1=FALSE)

Arguments

X
matrix of which the modes of the rows are estimated.

maxiter
maximal number of iterations; default = 50.

tol
tolerance for stopping; default = 0.001.

alpha
learning rate; default = 0.1.

a1
parameter of the width of the given distribution; default = 4.

G1
kind of distribution, TRUE: G1, FALSE: G2; default = FALSE.

Details

The mode is estimated by contrast functions G1

(1/a_1) * \ln (\cosh (a1*x))

or G2

- (1/a_1)*\exp(-1/2 * x*x)

The estimation is performed by gradient descent initialized by the median.

Implementation in R.

Value

u
the vector of estimated modes.

xu
X-u the mode centered data.

Author(s)

Sepp Hochreiter

References

A. Hyvaerinen, ‘Fast and Robust Fixed-Point Algorithms for Independent Component Analysis’, IEEE Transactions on Neural Networks 10(3):626-634, 1999.

See Also

fabia, fabias, fabiap, fabi, fabiasp, mfsc, nmfdiv, nmfeu, nmfsc, extractPlot, extractBic, plotBicluster, Factorization, projFuncPos, projFunc, estimateMode, makeFabiaData, makeFabiaDataBlocks, makeFabiaDataPos, makeFabiaDataBlocksPos, matrixImagePlot, fabiaDemo, fabiaVersion

Examples

#---------------
# DEMO
#---------------

dat <- makeFabiaDataBlocksPos(n = 100,l= 50,p = 10,f1 = 5,f2 = 5,
       of1 = 5,of2 = 10,sd_noise = 2.0,sd_z_noise = 0.2,mean_z = 2.0,
       sd_z = 1.0,sd_l_noise = 0.2,mean_l = 3.0,sd_l = 1.0)

X <- dat[[1]]

# modes <- estimateMode(X)

# modes$u - apply(X, 1, median)

[Package fabia version 2.38.0 Index]