readSamplesSpfabia {fabia}	R Documentation

Factor Analysis for Bicluster Acquisition: Read Sparse Matrix Samples

Description

readSamplesSpfabia: C implementation of readSamplesSpfabia.

Usage

readSamplesSpfabia(X,samples=0,lowerB=0.0,upperB=1000.0)

Arguments

X
the file name of the sparse matrix in sparse format.

samples
vector of samples which should be read; default = 0 (all samples)

lowerB
lower bound for filtering the inputs columns, the minimal column sum; default = 0.0.

upperB
upper bound for filtering the inputs columns, the maximal column sum; default = 1000.0.

Details

The data matrix is directly scanned by the C-code and must be in sparse matrix format.

Sparse matrix format: *first line: numer of rows (the samples). *second line: number of columns (the features). *following lines: for each sample (row) three lines with

I) number of nonzero row elements

II) indices of the nonzero row elements (ATTENTION: starts with 0!!)

III) values of the nonzero row elements (ATTENTION: floats with decimal point like 1.0 !!)

The code is implemented in C.

Value

X (data of the given samples)

Author(s)

Sepp Hochreiter

References

S. Hochreiter et al., ‘FABIA: Factor Analysis for Bicluster Acquisition’, Bioinformatics 26(12):1520-1527, 2010. http://bioinformatics.oxfordjournals.org/cgi/content/abstract/btq227

See Also

fabia, fabias, fabiap, spfabia, readSamplesSpfabia, readSpfabiaResult, fabi, fabiasp, mfsc, nmfdiv, nmfeu, nmfsc, extractPlot, extractBic, plotBicluster, Factorization, projFuncPos, projFunc, estimateMode, makeFabiaData, makeFabiaDataBlocks, makeFabiaDataPos, makeFabiaDataBlocksPos, matrixImagePlot, fabiaDemo, fabiaVersion

Examples

#---------------
# TEST
#---------------

[Package fabia version 2.38.0 Index]